Services + azure blobs + docker fixes in ProfilesAPI

+++ ProfilesAPI/ProfilesAPI.Services/Services/AdministratorService.cs
@@ -1,33 +1,44 @@
 ﻿using AutoMapper;
+using CommonLibrary.CommonService;
+using CommonLibrary.Constants;
 using FluentValidation;
 using InnoClinic.CommonLibrary.Exceptions;
 using InnoClinic.CommonLibrary.Response;
+using Microsoft.AspNetCore.Http;
 using ProfilesAPI.Domain.Data.Models;
 using ProfilesAPI.Domain.IRepositories;
 using ProfilesAPI.Services.Abstractions.Interfaces;
 using ProfilesAPI.Shared.DTOs.AdministratorDTOs;
+using ProfilesAPI.Shared.DTOs.DoctorDTOs;
+using System.Numerics;
 
 namespace ProfilesAPI.Services.Services;
 
 public class AdministratorService : IAdministratorService
 {
     private readonly IRepositoryManager _repositoryManager;
+    private readonly ICommonService _commonService;
+    private readonly IBlobStorageService _blobService;
     private readonly IMapper _mapper;
     private readonly IValidator<AdministratorForCreateDTO> _administratorForCreateValidator;
     private readonly IValidator<AdministratorForUpdateDTO> _administratorForUpdateValidator;
     public AdministratorService(
-            IRepositoryManager repositoryManager, 
-            IMapper mapper, 
-            IValidator<AdministratorForCreateDTO> administratorForCreateValidator, 
-            IValidator<AdministratorForUpdateDTO> administratorForUpdateValidator)
+            IRepositoryManager repositoryManager,
+            IMapper mapper,
+            IValidator<AdministratorForCreateDTO> administratorForCreateValidator,
+            IValidator<AdministratorForUpdateDTO> administratorForUpdateValidator,
+            ICommonService commonService,
+            IBlobStorageService blobService)
     {
         _repositoryManager = repositoryManager;
         _mapper = mapper;
         _administratorForCreateValidator = administratorForCreateValidator;
         _administratorForUpdateValidator = administratorForUpdateValidator;
+        _commonService = commonService;
+        _blobService = blobService;
     }
 
-    public async Task<ResponseMessage> AddAdministratorAsync(AdministratorForCreateDTO administratorForCreateDTO)
+    public async Task<ResponseMessage> AddAdministratorAsync(AdministratorForCreateDTO administratorForCreateDTO, IFormFile file)
     {
         var validationResult = await _administratorForCreateValidator.ValidateAsync(administratorForCreateDTO);
         if (!validationResult.IsValid)
@@ -35,7 +46,12 @@ public class AdministratorService : IAdministratorService
             throw new ValidationAppException(validationResult.Errors.Select(e => e.ErrorMessage).ToArray());
         }
 
+        var currentUserInfo = _commonService.GetCurrentUserInfo();
         var administrator = _mapper.Map<Administrator>(administratorForCreateDTO);
+        using Stream stream = file.OpenReadStream();
+        var fileId = await _blobService.UploadAsync(stream, file.ContentType);
+        administrator.UserId = currentUserInfo.Id;
+        administrator.Photo = fileId;
         await _repositoryManager.Administrator.AddAdministratorAsync(administrator);
 
         return new ResponseMessage();
@@ -43,6 +59,18 @@ public class AdministratorService : IAdministratorService
 
     public async Task<ResponseMessage> DeleteAdministratorByIdAsync(Guid administratorId)
     {
+        var administrator = await _repositoryManager.Administrator.GetAdministratorByIdAsync(administratorId);
+        if (administrator is null)
+        {
+            return new ResponseMessage("Administrator's Profile Not Found!", 404);
+        }
+
+        var currentUserInfo = _commonService.GetCurrentUserInfo();
+        if (currentUserInfo is null || !administrator.UserId.Equals(currentUserInfo.Id) || !currentUserInfo.Role.Equals(RoleConstants.Administrator)
+        {
+            return new ResponseMessage("Forbidden Action! You have no rights to manage this Administrator's Profile!", 403);
+        }
+
         await _repositoryManager.Administrator.DeleteAdministratorByIdAsync(administratorId);
 
         return new ResponseMessage();
@@ -74,7 +102,7 @@ public class AdministratorService : IAdministratorService
         return new ResponseMessage<ICollection<AdministratorTableInfoDTO>>(administratorTableInfoDTOs);
     }
 
-    public async Task<ResponseMessage> UpdateAdministratorAsync(Guid administratorId, AdministratorForUpdateDTO administratorForUpdateDTO)
+    public async Task<ResponseMessage> UpdateAdministratorAsync(Guid administratorId, AdministratorForUpdateDTO administratorForUpdateDTO, IFormFile? file)
     {
         var validationResult = await _administratorForUpdateValidator.ValidateAsync(administratorForUpdateDTO);
         if (!validationResult.IsValid)
@@ -88,7 +116,22 @@ public class AdministratorService : IAdministratorService
             return new ResponseMessage("Administrator's Profile Not Found!", 404);
         }
 
+        
+        var currentUserInfo = _commonService.GetCurrentUserInfo();
+        if (currentUserInfo is null || !administrator.UserId.Equals(currentUserInfo.Id))
+        {
+            return new ResponseMessage("Forbidden Action! You have no rights to manage this Administrator's Profile!", 403);
+        }
+
         administrator = _mapper.Map(administratorForUpdateDTO, administrator);
+        if (file is not null)
+        {
+            using Stream stream = file.OpenReadStream();
+            await _blobService.DeleteAsync(administrator.Photo);
+            var fileId = await _blobService.UploadAsync(stream, file.ContentType);
+            administrator.Photo = fileId;
+        }
+
         await _repositoryManager.Administrator.UpdateAdministratorAsync(administrator);
 
         return new ResponseMessage();
