+++ ProfilesAPI/ProfilesAPI.Services/Services/DoctorService.cs
@@ -1,17 +1,22 @@
 ﻿using AutoMapper;
+using CommonLibrary.CommonService;
+using CommonLibrary.Constants;
 using FluentValidation;
 using InnoClinic.CommonLibrary.Exceptions;
 using InnoClinic.CommonLibrary.Response;
+using Microsoft.AspNetCore.Http;
 using ProfilesAPI.Domain.Data.Models;
 using ProfilesAPI.Domain.IRepositories;
 using ProfilesAPI.Services.Abstractions.Interfaces;
 using ProfilesAPI.Shared.DTOs.DoctorDTOs;
 
 namespace ProfilesAPI.Services.Services;
-
+// think about specializations!!!!!
 public class DoctorService : IDoctorService
 {
     private readonly IRepositoryManager _repositoryManager;
+    private readonly IBlobStorageService _blobService;
+    private readonly ICommonService _commonService;
     private readonly IMapper _mapper;
     private readonly IValidator<DoctorForCreateDTO> _doctorForCreateValidator;
     private readonly IValidator<DoctorForUpdateDTO> _doctorForUpdateValidator;
@@ -19,16 +24,20 @@ public class DoctorService : IDoctorService
     public DoctorService(
             IRepositoryManager repositoryManager,
             IMapper mapper,
-            IValidator<DoctorForCreateDTO> doctorForCreateValidator, 
-            IValidator<DoctorForUpdateDTO> doctorForUpdateValidator)
+            IValidator<DoctorForCreateDTO> doctorForCreateValidator,
+            IValidator<DoctorForUpdateDTO> doctorForUpdateValidator,
+            ICommonService commonService,
+            IBlobStorageService blobService)
     {
         _repositoryManager = repositoryManager;
         _mapper = mapper;
         _doctorForCreateValidator = doctorForCreateValidator;
         _doctorForUpdateValidator = doctorForUpdateValidator;
+        _commonService = commonService;
+        _blobService = blobService;
     }
 
-    public async Task<ResponseMessage> AddDoctorAsync(DoctorForCreateDTO doctorForCreateDTO)
+    public async Task<ResponseMessage> AddDoctorAsync(DoctorForCreateDTO doctorForCreateDTO, IFormFile file)
     {
         var validationResult = await _doctorForCreateValidator.ValidateAsync(doctorForCreateDTO);
         if (!validationResult.IsValid)
@@ -36,7 +45,13 @@ public class DoctorService : IDoctorService
             throw new ValidationAppException(validationResult.Errors.Select(e => e.ErrorMessage).ToArray());
         }
 
+        var currentUserInfo = _commonService.GetCurrentUserInfo();
         var doctor = _mapper.Map<Doctor>(doctorForCreateDTO);
+        using Stream stream = file.OpenReadStream();
+        var fileId = await _blobService.UploadAsync(stream, file.ContentType);
+        doctor.UserId = currentUserInfo.Id;
+        doctor.Photo = fileId;
+
         await _repositoryManager.Doctor.AddDoctorAsync(doctor);
 
         return new ResponseMessage();
@@ -44,6 +59,18 @@ public class DoctorService : IDoctorService
 
     public async Task<ResponseMessage> DeleteDoctorByIdAsync(Guid doctorId)
     {
+        var doctor = await _repositoryManager.Doctor.GetDoctorByIdAsync(doctorId);
+        if(doctor is null)
+        {
+            return new ResponseMessage("Doctor's Profile Not Found!", 404);
+        }
+
+        var currentUserInfo = _commonService.GetCurrentUserInfo();
+        if (currentUserInfo is null || !doctor.UserId.Equals(currentUserInfo.Id) || !currentUserInfo.Role.Equals(RoleConstants.Administrator))
+        {
+            return new ResponseMessage("Forbidden Action! You have no rights to manage this Doctor's Profile!", 403);
+        }
+
         await _repositoryManager.Doctor.DeleteDoctorByIdAsync(doctorId);
 
         return new ResponseMessage();
@@ -75,7 +102,7 @@ public class DoctorService : IDoctorService
         return new ResponseMessage<DoctorInfoDTO>(doctorInfoDTO);
     }
 
-    public async Task<ResponseMessage> UpdateDoctorAsync(Guid doctorId, DoctorForUpdateDTO doctorForUpdateDTO)
+    public async Task<ResponseMessage> UpdateDoctorAsync(Guid doctorId, DoctorForUpdateDTO doctorForUpdateDTO, IFormFile? file)
     {
         var validationResult = await _doctorForUpdateValidator.ValidateAsync(doctorForUpdateDTO);
         if (!validationResult.IsValid)
@@ -89,7 +116,22 @@ public class DoctorService : IDoctorService
             return new ResponseMessage("Doctor's Profile Not Found!", 404);
         }
 
+
+        var currentUserInfo = _commonService.GetCurrentUserInfo();
+        if (currentUserInfo is null || !doctor.UserId.Equals(currentUserInfo.Id))
+        {
+            return new ResponseMessage("Forbidden Action! You have no rights to manage this Doctor's Profile!", 403);
+        }
+        
         doctor = _mapper.Map(doctorForUpdateDTO, doctor);
+        if (file is not null)
+        {
+            using Stream stream = file.OpenReadStream();
+            await _blobService.DeleteAsync(doctor.Photo);
+            var fileId = await _blobService.UploadAsync(stream, file.ContentType);
+            doctor.Photo = fileId;
+        }
+
         await _repositoryManager.Doctor.UpdateDoctorAsync(doctor);
 
         return new ResponseMessage();
